Handle EVTX files with malformed event XML
EVTX files that contain events with malformed string data generate a
UnicodeDecode exception, which halts all parsing of an otherwise valid
EVTX file. This fix silently ignores characters that cannot be decoded
as UTF-16 which seems reasonable considering that the data is supposed
to be string data.  Note: there may be a more appropriate place to
handle this scenario in a calling function, however, this fix has the
advantage of returning as much data as possible to the caller (rather
than halting processing).